Minibudget/pyproject.toml
[project]
name = "minibudget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
